Added more minimize options. Minimize profiles use a JSON format now to utilize these options.

+++ src/main/java/com/badlogicgames/packr/PackrReduce.java
@@ -16,14 +16,12 @@
 
 package com.badlogicgames.packr;
 
+import com.eclipsesource.json.*;
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.io.filefilter.TrueFileFilter;
 import org.zeroturnaround.zip.ZipUtil;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -40,118 +38,152 @@ class PackrReduce {
 		System.out.println("Minimizing JRE ...");
 
 		if (config.verbose) {
-			System.out.println("  # Unpacking rt.jar ...");
-		}
-
-		ZipUtil.unpack(new File(output, "jre/lib/rt.jar"), new File(output, "jre/lib/rt"));
-
-		// todo: detect if OracleJDK or OpenJDK
-		minimizeOpenJre(output, config);
-	}
-
-	private static void minimizeOpenJre(File output, PackrConfig config) throws IOException {
-		if (config.verbose) {
-			System.out.println("  # Processing OpenJDK runtime ...");
-		}
-
-		if (config.verbose) {
 			System.out.println("  # Removing executables ...");
 		}
 
-		if (config.platform == PackrConfig.Platform.Windows32 || config.platform == PackrConfig.Platform.Windows64) {
-			FileUtils.deleteDirectory(new File(output, "jre/bin/client"));
-			File[] files = new File(output, "jre/bin").listFiles();
-			if (files != null) {
-				for (File file : files) {
-					if (file.getName().endsWith(".exe")) {
-						PackrFileUtils.delete(file);
-					}
-				}
-			}
-		} else {
-			FileUtils.deleteDirectory(new File(output, "jre/bin"));
-		}
-
-		String[] minimizeList = readMinimizeProfile(config);
-		if (minimizeList.length > 0) {
+		JsonObject minimizeJson = readMinimizeProfile(config);
+		if (minimizeJson != null) {
 			if (config.verbose) {
 				System.out.println("  # Removing files and directories in profile '" + config.minimizeJre + "' ...");
 			}
 
-			for (String minimize : minimizeList) {
-				minimize = minimize.trim();
-				File file = new File(output, minimize);
-				try {
-					if (file.isDirectory()) {
-						FileUtils.deleteDirectory(new File(output, minimize));
-					} else {
-						PackrFileUtils.delete(file);
+			JsonArray reduceArray = minimizeJson.get("reduce").asArray();
+			for (JsonValue reduce : reduceArray) {
+				String path = reduce.asObject().get("archive").asString();
+				File file = new File(output, path);
+
+				if (!file.exists()) {
+					if (config.verbose) {
+						System.out.println("  # No file or directory '" + file.getPath() + "' found, skipping ...");
 					}
-				} catch (IOException e) {
-					System.out.println("Failed to delete " + file.getPath() + ": " + e.getMessage());
+					continue;
+				}
+
+				boolean needsUnpack = !file.isDirectory();
+
+				File fileNoExt = needsUnpack
+						? new File(output, path.contains(".") ? path.substring(0, path.lastIndexOf('.')) : path)
+						: file;
+
+				if (needsUnpack) {
+					if (config.verbose) {
+						System.out.println("  # Unpacking '" + file.getPath() + "' ...");
+					}
+					ZipUtil.unpack(file, fileNoExt);
+				}
+
+				JsonArray removeArray = reduce.asObject().get("paths").asArray();
+				for (JsonValue remove : removeArray) {
+					File removeFile = new File(fileNoExt, remove.asString());
+					if (removeFile.exists()) {
+						if (removeFile.isDirectory()) {
+							FileUtils.deleteDirectory(removeFile);
+						} else {
+							PackrFileUtils.delete(removeFile);
+						}
+					} else {
+						if (config.verbose) {
+							System.out.println("  # No file or directory '" + removeFile.getPath() + "' found");
+						}
+					}
+				}
+
+				if (needsUnpack) {
+					if (config.verbose) {
+						System.out.println("  # Repacking '" + file.getPath() + "' ...");
+					}
+
+					long beforeLen = file.length();
+					PackrFileUtils.delete(file);
+
+					ZipUtil.pack(fileNoExt, file);
+					FileUtils.deleteDirectory(fileNoExt);
+
+					long afterLen = file.length();
+
+					if (config.verbose) {
+						System.out.println("  # " + beforeLen / 1024 + " kb -> " + afterLen / 1024 + " kb");
+					}
+				}
+			}
+
+			JsonArray removeArray = minimizeJson.get("remove").asArray();
+			for (JsonValue remove : removeArray) {
+				String platform = remove.asObject().get("platform").asString();
+
+				if (!matchPlatformString(platform, config)) {
+					continue;
+				}
+
+				JsonArray removeFilesArray = remove.asObject().get("paths").asArray();
+				for (JsonValue removeFile : removeFilesArray) {
+					removeFileWildcard(output, removeFile.asString(), config);
 				}
 			}
 		}
+	}
 
-		if (new File(output, "jre/lib/rhino.jar").exists()) {
-			if (config.verbose) {
-				System.out.println("  # Removing rhino.jar ...");
+	private static boolean matchPlatformString(String platform, PackrConfig config) {
+		return "*".equals(platform) || config.platform.desc.contains(platform);
+	}
+
+	private static void removeFileWildcard(File output, String removeFileWildcard, PackrConfig config) throws IOException {
+		if (removeFileWildcard.contains("*")) {
+			String removePath = removeFileWildcard.substring(0, removeFileWildcard.indexOf('*') - 1);
+			String removeSuffix = removeFileWildcard.substring(removeFileWildcard.indexOf('*') + 1);
+
+			File[] files = new File(output, removePath).listFiles();
+			if (files != null) {
+				for (File file : files) {
+					if (removeSuffix.isEmpty() || file.getName().endsWith(removeSuffix)) {
+						removeFile(file, config);
+					}
+				}
+			} else {
+				if (config.verbose) {
+					System.out.println("  # No matching files found in '" + removeFileWildcard + "'");
+				}
 			}
-			PackrFileUtils.delete(new File(output, "jre/lib/rhino.jar"));
+		} else {
+			removeFile(new File(output, removeFileWildcard), config);
 		}
 	}
 
-	private static String[] readMinimizeProfile(PackrConfig config) throws IOException {
-		String[] lines;
+	private static void removeFile(File file, PackrConfig config) throws IOException {
+		if (!file.exists()) {
+			if (config.verbose) {
+				System.out.println("  # No file or directory '" + file.getPath() + "' found");
+			}
+			return;
+		}
+
+		if (config.verbose) {
+			System.out.println("  # Removing '" + file.getPath() + "'");
+		}
+
+		if (file.isDirectory()) {
+			FileUtils.deleteDirectory(file);
+		} else {
+			PackrFileUtils.delete(file);
+		}
+	}
+
+	private static JsonObject readMinimizeProfile(PackrConfig config) throws IOException {
+
+		JsonObject json;
 
 		if (new File(config.minimizeJre).exists()) {
-			lines = FileUtils.readFileToString(new File(config.minimizeJre)).split("\r?\n");
+			json = JsonObject.readFrom(FileUtils.readFileToString(new File(config.minimizeJre)));
 		} else {
 			InputStream in = Packr.class.getResourceAsStream("/minimize/" + config.minimizeJre);
 			if (in != null) {
-				lines = IOUtils.toString(in).split("\r?\n");
-				in.close();
+				json = JsonObject.readFrom(new InputStreamReader(in));
 			} else {
-				lines = new String[0];
+				json = null;
 			}
 		}
 
-		return lines;
-	}
-
-	static void repackJarFiles(File output, PackrConfig config) throws IOException {
-		repackJarFile(new File(output, "jre/lib/rt.jar"), new File(output, "jre/lib/rt"), config);
-	}
-
-	private static void repackJarFile(File jar, File fromDir, PackrConfig config) throws IOException {
-		System.out.println("Reducing " + jar.getName() + " ...");
-
-		if (fromDir.exists()) {
-			if (!fromDir.isDirectory()) {
-				throw new IOException("Expecting directory, but file found: " + fromDir.getPath());
-			}
-		} else {
-			if (config.verbose) {
-				System.out.println("  # Unpacking " + jar.getName() + " first ...");
-			}
-			ZipUtil.unpack(jar, fromDir);
-		}
-
-		long beforeLen = jar.length();
-		PackrFileUtils.delete(jar);
-
-		if (config.verbose) {
-			System.out.println("  # (Re-)packing " + jar.getName() + " ...");
-		}
-
-		ZipUtil.pack(fromDir, jar);
-		FileUtils.deleteDirectory(fromDir);
-
-		long afterLen = jar.length();
-
-		if (config.verbose) {
-			System.out.println("  # " + beforeLen / 1024 + " kb -> " + afterLen / 1024 + " kb");
-		}
+		return json;
 	}
 
 	static void removePlatformLibs(File output, PackrConfig config) throws IOException {
@@ -159,14 +191,20 @@ class PackrReduce {
 
 		// let's remove any shared libs not used on the platform, e.g. libGDX/LWJGL natives
 		for (String classpath : config.classpath) {
-			if (config.verbose) {
-				System.out.println("  # Unpacking '" + classpath + "'");
-			}
-
 			File jar = new File(output, new File(classpath).getName());
 			File jarDir = new File(output, jar.getName() + ".tmp");
 
-			ZipUtil.unpack(jar, jarDir);
+			if (config.verbose) {
+				if (jar.isDirectory()) {
+					System.out.println("  # Classpath '" + classpath + "' is a directory");
+				} else {
+					System.out.println("  # Unpacking '" + classpath + "' ...");
+				}
+			}
+
+			if (!jar.isDirectory()) {
+				ZipUtil.unpack(jar, jarDir);
+			}
 
 			Set<String> extensions = new HashSet<String>();
 
@@ -199,13 +237,22 @@ class PackrReduce {
 				}
 			}
 
-			if (config.verbose) {
-				System.out.println("  # Repacking '" + classpath + "'");
-			}
+			if (!jar.isDirectory()) {
+				if (config.verbose) {
+					System.out.println("  # Repacking '" + classpath + "' ...");
+				}
 
-			PackrFileUtils.delete(jar);
-			ZipUtil.pack(jarDir, jar);
-			FileUtils.deleteDirectory(jarDir);
+				long beforeLen = jar.length();
+				PackrFileUtils.delete(jar);
+
+				ZipUtil.pack(jarDir, jar);
+				FileUtils.deleteDirectory(jarDir);
+
+				long afterLen = jar.length();
+				if (config.verbose) {
+					System.out.println("  # " + beforeLen / 1024 + " kb -> " + afterLen / 1024 + " kb");
+				}
+			}
 		}
 	}
